Moved rest of framework to Int64

+++ FrameworkHs/Prims.hs
@@ -38,13 +38,14 @@ module FrameworkHs.Prims
   where
 
 import Prelude hiding (LT, EQ, GT)
+import Data.Int
 
 --------------------------------------------------------------------------------
 -- Terminal Type Definitions ---------------------------------------------------
 
-data UVar = UV String Int deriving (Read, Show, Eq)
-data FVar = FV Int        deriving (Read, Show, Eq)
-data Label = L String Int deriving (Read, Show, Eq)
+data UVar = UV String Int64 deriving (Read, Show, Eq)
+data FVar = FV Int64        deriving (Read, Show, Eq)
+data Label = L String Int64 deriving (Read, Show, Eq)
 
 data Reg = RAX | RCX | RDX | RBX | RBP | RSI | RDI | R8 | R9
          | R10 | R11 | R12 | R13 | R14 | R15
@@ -53,5 +54,5 @@ data Reg = RAX | RCX | RDX | RBX | RBP | RSI | RDI | R8 | R9
 data Relop = LT | LTE | EQ | GT | GTE               deriving (Read,Show,Eq,Ord)
 data Binop = MUL | ADD | SUB | LOGAND | LOGOR | SRA deriving (Read,Show,Eq,Ord)
 
-data Disp = Disp Reg Int
+data Disp = Disp Reg Int64
 data Ind = Ind Reg Reg
